moved shacl code to own package
