+++ backend/molgenis-emx2-datamodels/src/test/java/org/molgenis/emx2/datamodels/RunRDFExportAndCheckCompliance.java
@@ -1,6 +1,6 @@
 package org.molgenis.emx2.datamodels;
 
-import static org.molgenis.emx2.rdf.SHACLComplianceTester.*;
+import static org.molgenis.emx2.rdf.shacl.SHACLComplianceTester.*;
 
 import java.io.File;
 import java.io.PrintWriter;
