+++ backend/molgenis-emx2-rdf/src/main/java/org/molgenis/emx2/rdf/shacl/SHACLComplianceTester.java
@@ -1,4 +1,4 @@
-package org.molgenis.emx2.rdf;
+package org.molgenis.emx2.rdf.shacl;
 
 public class SHACLComplianceTester {
   /**
